The Password Pastor - All Files
THIS CODE WILL NOT WORK ON YOUR DEVICE.

This code uses a MySQL database called PasswordPastor that I run off Port 3306 of my computer. This project was programmed for my Computer Science A Level project which was scored an A. If you read the source code, you'll be able to view some of my experience with SQL and Python in an Object Oriented Format. Please contact me if you would like to know more about this project.